% Generated by build script
\name{doj_ch7_data}
\alias{doj_ch7_data}
\title{Get Chapter 7 Trustee Final Report CSV data}
\usage{
doj_ch7_data(year = 2016)
}
\arguments{
\item{year}{Calendar year (e.g. 2016). Defaults to 2016 (latest CSV).}
}
\value{tibble of Chapter 7 trustee final report data}
\description{Downloads and parses a specific Chapter 7 CSV data file.}
