% Generated by build script
\name{doj_it_leadership}
\alias{doj_it_leadership}
\title{Get DOJ IT Leadership (Bureau Directory)}
\usage{
doj_it_leadership()
}
\value{tibble with bureau_code, first_name, last_name, employment_type, appointment_type, email, component, title, key_bureau_cio}
\description{Returns the Department of Justice bureau CIO directory as a tibble.}
